Moving logic and rename classes
